+++ libs/i18n/src/lib/translations/translations.ts
@@ -532,7 +532,8 @@ export default {
         confirmBtnLabel: 'OK',
         cancelBtnLabel: 'Cancel',
         filteredBy: 'Filtered by',
-        showingBlankOfBlank: 'Showing {totalGrowingItemsCount} of {totalGrowingItemsCount}'
+        showingBlankOfBlank: 'Showing {totalGrowingItemsCount} of {totalGrowingItemsCount}',
+        clearFilters: 'Clear Filters'
     },
     platformWizardGenerator: {
         summarySectionEditStep: 'Edit'