+++ libs/platform/form/multi-combobox/commons/base-multi-combobox.ts
@@ -571,7 +571,10 @@ export abstract class BaseMultiCombobox extends CollectionBaseInput implements O
             this.toggleSelectionByInputText();
         } else if (!KeyUtil.isKeyCode(event, [...this._nonOpeningKeys, CONTROL])) {
             this.showList(true);
-            const acceptedKeys = !KeyUtil.isKeyType(event, 'alphabetical') && !KeyUtil.isKeyType(event, 'numeric');
+            const acceptedKeys =
+                !KeyUtil.isKeyType(event, 'alphabetical') &&
+                !KeyUtil.isKeyType(event, 'numeric') &&
+                !KeyUtil.isKeyType(event, 'ime');
             if (acceptedKeys) {
                 // SetTimeout is needed for input to receive new value.
                 setTimeout(() => {